:truck: [https://github.com/open-formulieren/open-forms/issues/2471] separate appointment components into several files
